+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [2.11.5]
+- Improve Keyboard Controls.
+- Adds support to tileset with individual image
+
 # [2.11.4]
 - Fix exception in `TiledWorldBuilder`.